@@ -22,8 +22,8 @@
   ],
   "license": "MIT",
   "peerDependencies": {
-    "react": "*",
-    "react-router": "^2.0.1"
+    "react": "^15.5.4",
+    "react-router": "^3.2.0"
   },
   "devDependencies": {
     "babel-cli": "^6.6.5",
@@ -33,6 +33,7 @@
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-1": "^6.5.0",
+    "create-react-class": "15.5.4",
     "eslint": "1.4.1",
     "eslint-config-rackt": "1.0.0",
     "eslint-plugin-react": "3.3.2",
@@ -51,9 +52,10 @@
     "karma-webpack": "^1.7.0",
     "mocha": "^2.0.1",
     "pretty-bytes": "^2.0.1",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "react-router": "^2.0.1",
+    "prop-types": "^15.6.0",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
+    "react-router": "^3.2.0",
     "rimraf": "^2.4.2",
     "webpack": "^1.12.6",
     "webpack-dev-middleware": "^1.2.0"
